+++ HISTORY.rst
@@ -33,6 +33,7 @@ End-User Summary
 - In priorization mode, ORPHA and DECIPHER terms are now selectable.
 - Fixed bug of wrong order when sorting by LOEUF score.
 - Adding some UI documenation.
+- Fixed bug where case alignment stats were not properly imported.
 
 Full Change List
 ================
@@ -72,6 +73,8 @@ Full Change List
 - In priorization mode, ORPHA and DECIPHER terms are now selectable.
 - Fixed bug of wrong order when sorting by LOEUF score.
 - Adding some UI documenation.
+- Fixed bug where case alignment stats were not properly imported.
+  Refactored case import in a sense that the new variant set gets activated when it is successfully imported.
 
 -------
 v0.21.0